Various changes to handle the SuperTag Field
git-svn-id: http://opensource.washingtontimes.com/projects/public/django-supertagging/trunk@33 a3a5eddf-6203-49af-8a4a-b68c2cc3a614

+++ supertagging/modules/__init__.py
@@ -6,6 +6,7 @@ import re
 from django.conf import settings
 from django.contrib.contenttypes.models import ContentType
 from django.template.defaultfilters import slugify
+from django.utils.encoding import force_unicode
 
 try:
     from calais import Calais
@@ -14,15 +15,28 @@ except ImportError:
 
 from supertagging import settings
 from supertagging.models import SuperTag, SuperTagRelation, SuperTaggedItem, SuperTaggedRelationItem
+from supertagging.utils import unicode_to_ascii
 
 REF_REGEX = "^http://d.opencalais.com/(?P<key>.*)$"
 
-def process(field, data, obj, process_type='TEXT/RAW'):
+def process(obj, tags=[]):
     """
     Process the data.
     """
     if not Calais:
-        raise "python-calais module was not found."
+        raise ImportError("python-calais module was not found.")
+    if not settings.API_KEY:
+        raise ValueError('Calais API KEY is missing.')
+        
+    params = settings.MODULES['%s.%s' % (obj._meta.app_label, obj._meta.module_name)]
+    
+    process_type = settings.DEFAULT_PROCESS_TYPE
+    if 'contentType' in settings.PROCESSING_DIR:
+        d_proc_type = proc_dir['contentType']
+        
+    if 'fields' not in params:
+        raise Exception('No "fields" found.')
+        
     # Create the instance of Calais and setup the parameters,
     # see open-calais.com for more information about user directives,
     # and processing directives
@@ -30,24 +44,48 @@ def process(field, data, obj, process_type='TEXT/RAW'):
     c.user_directives.update(settings.USER_DIR)
     c.processing_directives.update(settings.PROCESSING_DIR)
     c.processing_directives['contentType'] = process_type
-    # Analyze the text (data)
-    result = c.analyze(data)
     
-    # Retrieve the Django content type for the obj
-    ctype = ContentType.objects.get_for_model(obj)
-    # Remove existing items, this ensures tagged items are updated correctly
-    SuperTaggedItem.objects.filter(content_type=ctype, object_id=obj.pk, field=field).delete()
-    SuperTaggedRelationItem.objects.filter(content_type=ctype, object_id=obj.pk, field=field).delete()
+    processed_tags = []
+    for item in params['fields']:
+        d = item.copy()
+        field = d.pop('name')
+        proc_type = d.pop('process_type', process_type)
+        
+        data = getattr(obj, field)
+        
+        #if isinstance(data, unicode):
+        #    data = unicode_to_ascii(data)
+        #else:
+        #    data = unicode_to_ascii(unicode(data, 'utf-8'))
+
+            
+        data = force_unicode(getattr(obj, field))
+        
+        # Analyze the text (data)
+        result = c.analyze(data)
     
-    # Process entities, relations and topics
-    if hasattr(result, 'entities'):
-        entities = _processEntities(field, result.entities, obj, ctype, process_type)
+        # Retrieve the Django content type for the obj
+        ctype = ContentType.objects.get_for_model(obj)
+        # Remove existing items, this ensures tagged items are updated correctly
+        SuperTaggedItem.objects.filter(content_type=ctype, object_id=obj.pk, field=field).delete()
+        if settings.PROCESS_RELATIONS:
+            SuperTaggedRelationItem.objects.filter(content_type=ctype, object_id=obj.pk, field=field).delete()
+    
+        entities, relations, topics = [], [], []
+        # Process entities, relations and topics
+        if hasattr(result, 'entities'):
+            entities = _processEntities(field, result.entities, obj, ctype, proc_type, tags)
         
-    if hasattr(result, 'relations') and settings.PROCESS_RELATIONS:
-        relations = _processRelations(field, result.relations, obj, ctype, process_type)
+        if hasattr(result, 'relations') and settings.PROCESS_RELATIONS:
+            relations = _processRelations(field, result.relations, obj, ctype, proc_type, tags)
         
-    if hasattr(result, 'topics') and settings.PROCESS_TOPICS:
-        topics =  _processTopics(field, result.topics, obj, ctype)
+        if hasattr(result, 'topics') and settings.PROCESS_TOPICS:
+            topics =  _processTopics(field, result.topics, obj, ctype, tags)
+        
+        processed_tags.extend(entities)
+        processed_tags.extend(topics)
+        
+    return processed_tags
     
 def clean_up(obj):
     """
@@ -60,10 +98,11 @@ def clean_up(obj):
     # TODO, clean up tags that have no related items?
     # Same for relations?
     
-def _processEntities(field, data, obj, ctype, process_type):
+def _processEntities(field, data, obj, ctype, process_type, tags):
     """
     Process Entities.
     """
+    processed_tags = []
     for entity in data:
         # Here we convert the given float value to an integer
         rel = int(float(str(entity.pop('relevance', '0'))) * 1000)
@@ -74,27 +113,34 @@ def _processEntities(field, data, obj, ctype, process_type):
         
         if stype.lower() not in settings.EXCLUSIONS:
             name = entity.pop('name', '').lower()
+            if tags and name not in tags:
+                continue
+                
             slug = slugify(name)
             try:
                 tag = SuperTag.objects.get(pk=pk)
             except SuperTag.DoesNotExist:
                 tag = SuperTag.objects.create(id=pk, slug=slug, stype=stype, name=name)
-        
+    
             tag.properties = entity
-            tag.save() 
+            tag.save()
+            
             #TODO: check to make sure that the entity is not already attached
             # to the content object, if it is, just append the instances. This
             # should elimiate entities returned with different names such as 
             # 'Washington' and 'Washington DC' but same id
             try:
-                mit = SuperTaggedItem.objects.get(tag=tag, content_type=ctype, object_id=obj.pk, field=field)
-                mit.instances.append(inst)
-                mit.save()
+                it = SuperTaggedItem.objects.get(tag=tag, content_type=ctype, object_id=obj.pk, field=field)
+                it.instances.append(inst)
+                it.save()
             except SuperTaggedItem.DoesNotExist:
                 # Create the record that will associate content to tags
                 it = SuperTaggedItem.objects.create(tag=tag, content_type=ctype, object_id=obj.pk, field=field, process_type=process_type, relevance=rel, instances=inst)
-        
-def _processRelations(field, data, obj, ctype, process_type):
+    
+            processed_tags.append(tag)
+    return processed_tags
+    
+def _processRelations(field, data, obj, ctype, process_type, tags):
     """
     Process Relations
     """
@@ -127,22 +173,29 @@ def _processRelations(field, data, obj, ctype, process_type):
                     
             _vals.update(props)
             
+            if tags and entity_value not in tags:
+                continue
+        
             entity = SuperTag.objects.get(pk=entity_value)
             rel_item, rel_created = SuperTagRelation.objects.get_or_create(tag=entity, name=entity_key, stype=rel_type, properties=_vals)
             
             SuperTaggedRelationItem.objects.create(relation=rel_item, content_type=ctype, object_id=obj.pk, field=field, process_type=process_type, instances=inst) 
             
-def _processTopics(field, data, obj, ctype):
+def _processTopics(field, data, obj, ctype, tags):
     """
     Process Topics, this opertaion is similar to _processEntities, the only
     difference is that there are no instances
     """
+    processed_tags = []
     for di in data:
         di.pop('__reference')
         
         pk = re.match(REF_REGEX, str(di.pop('category'))).group('key')
         stype = 'Topic'
         name = di.pop('categoryName', '').lower()
+        if tags and name not in tags:
+            continue
+            
         slug = slugify(name)
         try:
             tag = SuperTag.objects.get(pk=pk)
@@ -154,6 +207,9 @@ def _processTopics(field, data, obj, ctype):
         
         SuperTaggedItem.objects.create(tag=tag, content_type=ctype, object_id=obj.pk, field=field)
         
+        processed_tags.append(tag)
+    return processed_tags
+    
 def _getEntityText(key):
     """
     Try to resolve the entity given the key
